requirements.txt
matplotlib~=3.3.1
numpy~=1.19.1
torchvision~=0.7.0
parse~=1.18.0
scikit-image~=0.16.2
pillow~=8.0.0
torch~=1.6.0